demos/dbt_demo/dbt_project/pyproject.toml
[tool.poetry]
name = "dbt_project"
version = "0.1.0"
description = ""
authors = ["Flavio Malavazi <flavio.malavazi1@gmail.com>"]
license = "GNU"

[tool.poetry.dependencies]
python = "3.8.6"
dbt-databricks = "^1.7.2"
apache-airflow = "^2.7.0"
apache-airflow-providers-databricks = "^6.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
